# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EventHub API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  icon: string;
  eventCount: number;
}

export interface Event {
  id: number;
  title: string;
  description: string;
  shortDescription: string;
  imageUrl?: string;
  startDate: string;
  endDate: string;
  location: string;
  venue: string;
  city: string;
  state?: string;
  isOnline: boolean;
  isFeatured: boolean;
  isFree: boolean;
  minPrice?: number;
  maxPrice?: number;
  categoryId: number;
  categoryName: string;
  organizerName: string;
  attendeeCount: number;
  totalTickets: number;
  availableTickets: number;
  createdAt: string;
}

export interface EventListResponse {
  events: Event[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface CreateEventBody {
  title: string;
  description: string;
  shortDescription: string;
  imageUrl?: string;
  startDate: string;
  endDate: string;
  location: string;
  venue: string;
  city: string;
  state?: string;
  isOnline: boolean;
  isFree: boolean;
  categoryId: number;
  organizerName: string;
}

export interface Ticket {
  id: number;
  eventId: number;
  name: string;
  description?: string;
  price: number;
  quantity: number;
  quantitySold: number;
  isFree: boolean;
}

export interface Registration {
  id: number;
  eventId: number;
  eventTitle: string;
  ticketId: number;
  ticketName: string;
  attendeeName: string;
  attendeeEmail: string;
  quantity: number;
  totalAmount: number;
  status: string;
  confirmationCode: string;
  createdAt: string;
}

export interface CreateRegistrationBody {
  eventId: number;
  ticketId: number;
  attendeeName: string;
  attendeeEmail: string;
  quantity: number;
}

export type ListEventsParams = {
  category?: string;
  search?: string;
  location?: string;
  dateFrom?: string;
  dateTo?: string;
  isFree?: boolean;
  page?: number;
  limit?: number;
};

export type ListTicketsByEventParams = {
  eventId: number;
};