# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EventHub API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all events
 */
export const listEventsQueryPageDefault = 1;
export const listEventsQueryLimitDefault = 12;

export const ListEventsQueryParams = zod.object({
  category: zod.coerce.string().optional(),
  search: zod.coerce.string().optional(),
  location: zod.coerce.string().optional(),
  dateFrom: zod.coerce.string().optional(),
  dateTo: zod.coerce.string().optional(),
  isFree: zod.coerce.boolean().optional(),
  page: zod.coerce.number().default(listEventsQueryPageDefault),
  limit: zod.coerce.number().default(listEventsQueryLimitDefault),
});

export const ListEventsResponse = zod.object({
  events: zod.array(
    zod.object({
      id: zod.number(),
      title: zod.string(),
      description: zod.string(),
      shortDescription: zod.string(),
      imageUrl: zod.string().optional(),
      startDate: zod.string(),
      endDate: zod.string(),
      location: zod.string(),
      venue: zod.string(),
      city: zod.string(),
      state: zod.string().optional(),
      isOnline: zod.boolean(),
      isFeatured: zod.boolean(),
      isFree: zod.boolean(),
      minPrice: zod.number().optional(),
      maxPrice: zod.number().optional(),
      categoryId: zod.number(),
      categoryName: zod.string(),
      organizerName: zod.string(),
      attendeeCount: zod.number(),
      totalTickets: zod.number(),
      availableTickets: zod.number(),
      createdAt: zod.string(),
    }),
  ),
  total: zod.number(),
  page: zod.number(),
  limit: zod.number(),
  totalPages: zod.number(),
});

/**
 * @summary Create a new event
 */
export const CreateEventBody = zod.object({
  title: zod.string(),
  description: zod.string(),
  shortDescription: zod.string(),
  imageUrl: zod.string().optional(),
  startDate: zod.string(),
  endDate: zod.string(),
  location: zod.string(),
  venue: zod.string(),
  city: zod.string(),
  state: zod.string().optional(),
  isOnline: zod.boolean(),
  isFree: zod.boolean(),
  categoryId: zod.number(),
  organizerName: zod.string(),
});

/**
 * @summary Get featured events
 */
export const GetFeaturedEventsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string(),
  shortDescription: zod.string(),
  imageUrl: zod.string().optional(),
  startDate: zod.string(),
  endDate: zod.string(),
  location: zod.string(),
  venue: zod.string(),
  city: zod.string(),
  state: zod.string().optional(),
  isOnline: zod.boolean(),
  isFeatured: zod.boolean(),
  isFree: zod.boolean(),
  minPrice: zod.number().optional(),
  maxPrice: zod.number().optional(),
  categoryId: zod.number(),
  categoryName: zod.string(),
  organizerName: zod.string(),
  attendeeCount: zod.number(),
  totalTickets: zod.number(),
  availableTickets: zod.number(),
  createdAt: zod.string(),
});
export const GetFeaturedEventsResponse = zod.array(
  GetFeaturedEventsResponseItem,
);

/**
 * @summary Get event by ID
 */
export const GetEventParams = zod.object({
  id: zod.coerce.number(),
});

export const GetEventResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string(),
  shortDescription: zod.string(),
  imageUrl: zod.string().optional(),
  startDate: zod.string(),
  endDate: zod.string(),
  location: zod.string(),
  venue: zod.string(),
  city: zod.string(),
  state: zod.string().optional(),
  isOnline: zod.boolean(),
  isFeatured: zod.boolean(),
  isFree: zod.boolean(),
  minPrice: zod.number().optional(),
  maxPrice: zod.number().optional(),
  categoryId: zod.number(),
  categoryName: zod.string(),
  organizerName: zod.string(),
  attendeeCount: zod.number(),
  totalTickets: zod.number(),
  availableTickets: zod.number(),
  createdAt: zod.string(),
});

/**
 * @summary List all categories
 */
export const ListCategoriesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  slug: zod.string(),
  icon: zod.string(),
  eventCount: zod.number(),
});
export const ListCategoriesResponse = zod.array(ListCategoriesResponseItem);

/**
 * @summary List tickets for an event
 */
export const ListTicketsByEventQueryParams = zod.object({
  eventId: zod.coerce.number(),
});

export const ListTicketsByEventResponseItem = zod.object({
  id: zod.number(),
  eventId: zod.number(),
  name: zod.string(),
  description: zod.string().optional(),
  price: zod.number(),
  quantity: zod.number(),
  quantitySold: zod.number(),
  isFree: zod.boolean(),
});
export const ListTicketsByEventResponse = zod.array(
  ListTicketsByEventResponseItem,
);

/**
 * @summary List all registrations
 */
export const ListRegistrationsResponseItem = zod.object({
  id: zod.number(),
  eventId: zod.number(),
  eventTitle: zod.string(),
  ticketId: zod.number(),
  ticketName: zod.string(),
  attendeeName: zod.string(),
  attendeeEmail: zod.string(),
  quantity: zod.number(),
  totalAmount: zod.number(),
  status: zod.string(),
  confirmationCode: zod.string(),
  createdAt: zod.string(),
});
export const ListRegistrationsResponse = zod.array(
  ListRegistrationsResponseItem,
);

/**
 * @summary Register for an event / purchase tickets
 */
export const CreateRegistrationBody = zod.object({
  eventId: zod.number(),
  ticketId: zod.number(),
  attendeeName: zod.string(),
  attendeeEmail: zod.string(),
  quantity: zod.number(),
});