# lib/api-zod/src/generated/types/createEventBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EventHub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateEventBody {
  title: string;
  description: string;
  shortDescription: string;
  imageUrl?: string;
  startDate: string;
  endDate: string;
  location: string;
  venue: string;
  city: string;
  state?: string;
  isOnline: boolean;
  isFree: boolean;
  categoryId: number;
  organizerName: string;
}