# lib/api-zod/src/generated/types/event.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EventHub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Event {
  id: number;
  title: string;
  description: string;
  shortDescription: string;
  imageUrl?: string;
  startDate: string;
  endDate: string;
  location: string;
  venue: string;
  city: string;
  state?: string;
  isOnline: boolean;
  isFeatured: boolean;
  isFree: boolean;
  minPrice?: number;
  maxPrice?: number;
  categoryId: number;
  categoryName: string;
  organizerName: string;
  attendeeCount: number;
  totalTickets: number;
  availableTickets: number;
  createdAt: string;
}